Support inline HTML in markdown documents
Since this is for creating a static site, there is no notable risk of
dangerous HTML being present in the markdown files. This allows for much
greater flexibility when writing markdown files, and fits nicely with
the no-config goal of ssg since it allows users to go outside the bounds
of markdown without having to configure anything; they can simply add
some HTML to their document.